stats/codes_commons, stats/period_commons: split stats commons
