+++ lib/3scale/backend/stats/period_commons.rb
@@ -1,7 +1,7 @@
 module ThreeScale
   module Backend
     module Stats
-      module Commons
+      module PeriodCommons
         SERVICE_GRANULARITIES = %i[eternity month week day hour].map do |g|
           Period[g]
         end.freeze
@@ -27,16 +27,6 @@ module ThreeScale
 
         def self.expire_time_for_granularity(granularity)
           GRANULARITY_EXPIRATION_TIME[granularity]
-        end
-
-        TRACKED_CODES = [200, 404, 403, 500, 503].freeze
-
-        HTTP_CODE_GROUPS_MAP = TRACKED_CODES.each_with_object({}) do |code, hsh|
-          hsh[code / 100] = "#{code / 100}XX"
-        end
-
-        def self.get_http_code_group(http_code)
-          HTTP_CODE_GROUPS_MAP.fetch(http_code / 100)
         end
       end
     end
